refact: rename plugin
