+++ lua/inline/init.lua
@@ -1,11 +1,11 @@
 local M = {}
 
 local function setup_commands()
-  local ln = require('line_notes')
+  local il = require('inline')
   -- Set up autocmd for buffer events
   vim.api.nvim_create_autocmd({ 'BufReadPost', 'BufEnter' }, {
     callback = function()
-      ln.notes.load_for_buffer()
+      il.notes.load_for_buffer()
     end,
   })
   -- easy escape from float
@@ -18,11 +18,11 @@ local function setup_commands()
     end,
   })
 
-  vim.api.nvim_create_user_command('LineNotes', function(opts)
+  vim.api.nvim_create_user_command('InlineNotes', function(opts)
     local args = vim.split(opts.args, '%s+', { trimempty = true })
     -- Default
     if #args == 0 then
-      ln.notes.show()
+      il.notes.show()
       return
     end
 
@@ -30,19 +30,19 @@ local function setup_commands()
     table.remove(args, 1) -- Remove command
 
     if command == 'show' then
-      ln.notes.show()
+      il.notes.show()
     elseif command == 'edit' then
-      ln.notes.show()
+      il.notes.show()
     elseif command == 'file' then
-      ln.notes.edit_file_note()
+      il.notes.edit_file_note()
     elseif command == 'add' then
-      ln.notes.add()
+      il.notes.add()
     elseif command == 'move' then
-      ln.notes.move()
+      il.notes.move()
     elseif command == 'delete' then
-      ln.notes.delete()
+      il.notes.delete()
     elseif command == 'search' then
-      ln.telescope_picker.open_notes_picker()
+      il.telescope_picker.open_notes_picker()
     end
   end, {
     nargs = '*',
@@ -66,8 +66,8 @@ end
 
 -- Define signs based on configuration
 local function setup_signs()
-  local sign_config = require('line_notes.config').config.signcolumn
-  vim.fn.sign_define('LineNote', {
+  local sign_config = require('inline.config').config.signcolumn
+  vim.fn.sign_define('InlineNote', {
     text = sign_config.note_icon,
     texthl = sign_config.highlight,
     numhl = sign_config.number_highlight,
@@ -75,20 +75,20 @@ local function setup_signs()
 end
 
 local function setup_keymaps(keymaps)
-  local ln = require('line_notes')
+  local il = require('inline')
   local function map(key, cmd)
     vim.keymap.set('n', key, cmd, { noremap = true, silent = true })
   end
 
-  map(keymaps.add_note, ln.notes.add_note)
-  map(keymaps.delete_note, ln.notes.delete_note)
-  map(keymaps.show_note, ln.notes.show_note)
-  map(keymaps.list_notes, ln.telescope_picker.open_notes_picker)
+  map(keymaps.add_note, il.notes.add_note)
+  map(keymaps.delete_note, il.notes.delete_note)
+  map(keymaps.show_note, il.notes.show_note)
+  map(keymaps.list_notes, il.telescope_picker.open_notes_picker)
 end
 
 function M.setup(opts)
   -- Initialize configuration
-  local config = require('line_notes.config').setup(opts)
+  local config = require('inline.config').setup(opts)
 
   if config.keymaps.enabled then
     setup_keymaps(config.keymaps)
@@ -97,12 +97,12 @@ function M.setup(opts)
   setup_signs()
   vim.treesitter.language.register('markdown', 'line_notes')
   -- Load notes
-  require('line_notes').notes.load()
+  require('inline').notes.load()
   return M
 end
 
-M.config = require('line_notes.config')
-M.notes = require('line_notes.notes')
-M.telescope_picker = require('line_notes.telescope_picker')
+M.config = require('inline.config')
+M.notes = require('inline.notes')
+M.telescope_picker = require('inline.telescope_picker')
 
 return M
