+++ plugin/inline.vim
@@ -1,0 +1,4 @@
+if exists('g:loaded_inline') | finish | endif
+let g:loaded_inline = 1
+
+" lua require('inline').setup()
