--- plugin/line_notes.vim
@@ -1,4 +1,0 @@
-if exists('g:loaded_line_notes') | finish | endif
-let g:loaded_line_notes = 1
-
-" lua require('line_notes').setup()
